r4efactor: extract report process listener to a file
